+++ packages/Webkul/Product/src/Http/Controllers/ProductController.php
@@ -7,12 +7,11 @@ use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Event;
 use Webkul\Product\Http\Requests\ProductForm;
 use Webkul\Product\Repositories\ProductRepository as Product;
-use Webkul\Product\Repositories\ProductGridRepository as ProductGrid;
-use Webkul\Product\Repositories\ProductFlatRepository as ProductFlat;
 use Webkul\Product\Repositories\ProductAttributeValueRepository as ProductAttributeValue;
 use Webkul\Attribute\Repositories\AttributeFamilyRepository as AttributeFamily;
 use Webkul\Category\Repositories\CategoryRepository as Category;
 use Webkul\Inventory\Repositories\InventorySourceRepository as InventorySource;
+use Illuminate\Support\Facades\Storage;
 
 /**
  * Product controller
@@ -58,21 +57,20 @@ class ProductController extends Controller
     protected $product;
 
     /**
-     * ProductGrid Repository object
+     * ProductAttributeValueRepository object
      *
      * @var array
      */
-    protected $productGrid;
-    protected $productFlat;
     protected $productAttributeValue;
 
     /**
      * Create a new controller instance.
      *
-     * @param  Webkul\Attribute\Repositories\AttributeFamilyRepository  $attributeFamily
-     * @param  Webkul\Category\Repositories\CategoryRepository          $category
-     * @param  Webkul\Inventory\Repositories\InventorySourceRepository  $inventorySource
-     * @param  Webkul\Product\Repositories\ProductRepository            $product
+     * @param  \Webkul\Attribute\Repositories\AttributeFamilyRepository     $attributeFamily
+     * @param  \Webkul\Category\Repositories\CategoryRepository             $category
+     * @param  \Webkul\Inventory\Repositories\InventorySourceRepository     $inventorySource
+     * @param  \Webkul\Product\Repositories\ProductRepository               $product
+     * @param  \Webkul\Product\Repositories\ProductAttributeValueRepository $productAttributeValue
      * @return void
      */
     public function __construct(
@@ -80,9 +78,8 @@ class ProductController extends Controller
         Category $category,
         InventorySource $inventorySource,
         Product $product,
-        ProductGrid $productGrid,
-        ProductFlat $productFlat,
-        ProductAttributeValue $productAttributeValue)
+        ProductAttributeValue $productAttributeValue
+    )
     {
         $this->attributeFamily = $attributeFamily;
 
@@ -91,10 +88,6 @@ class ProductController extends Controller
         $this->inventorySource = $inventorySource;
 
         $this->product = $product;
-
-        $this->productGrid = $productGrid;
-
-        $this->productFlat = $productFlat;
 
         $this->productAttributeValue = $productAttributeValue;
 
@@ -167,15 +160,13 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
-        $product = $this->product->with(['variants'])->find($id);
+        $product = $this->product->with(['variants', 'variants.inventories'])->findOrFail($id);
 
         $categories = $this->category->getCategoryTree();
 
         $inventorySources = $this->inventorySource->all();
 
-        $allProducts = $this->productGrid->all();
-
-        return view($this->_config['view'], compact('product', 'categories', 'inventorySources', 'allProducts'));
+        return view($this->_config['view'], compact('product', 'categories', 'inventorySources'));
     }
 
     /**
@@ -202,11 +193,19 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        $this->product->delete($id);
+        $product = $this->product->findOrFail($id);
 
-        session()->flash('success', trans('admin::app.response.delete-success', ['name' => 'Product']));
+        try {
+            $this->product->delete($id);
 
-        return redirect()->back();
+            session()->flash('success', trans('admin::app.response.delete-success', ['name' => 'Product']));
+
+            return response()->json(['message' => true], 200);
+        } catch (\Exception $e) {
+            session()->flash('error', trans('admin::app.response.delete-failed', ['name' => 'Product']));
+        }
+
+        return response()->json(['message' => false], 400);
     }
 
     /**
@@ -219,7 +218,11 @@ class ProductController extends Controller
         $productIds = explode(',', request()->input('indexes'));
 
         foreach ($productIds as $productId) {
-            $this->product->delete($productId);
+            $product = $this->product->find($productId);
+
+            if (isset($product)) {
+                $this->product->delete($productId);
+            }
         }
 
         session()->flash('success', trans('admin::app.catalog.products.mass-delete-success'));
@@ -291,5 +294,21 @@ class ProductController extends Controller
         } else {
             return view($this->_config['view']);
         }
+    }
+
+     /**
+     * Download image or file
+     *
+     * @param  int $productId, $attributeId
+     * @return \Illuminate\Http\Response
+     */
+    public function download($productId, $attributeId)
+    {
+        $productAttribute = $this->productAttributeValue->findOneWhere([
+            'product_id'   => $productId,
+            'attribute_id' => $attributeId
+        ]);
+
+        return Storage::download($productAttribute['text_value']);
     }
 }
